Use lazytensors

+++ py-lddmm/Tests/surfaceMatchingTest.py
@@ -183,8 +183,8 @@ def compute(model):
 
     sm = SurfaceMatchingParam(timeStep=0.1, algorithm='bfgs', KparDiff=K1, KparDist=('gauss', sigmaDist),
                               sigmaError=sigmaError, errorType='varifold', internalCost=internalCost)
-    sm.KparDiff.pk_dtype = 'float32'
-    sm.KparDist.pk_dtype = 'float32'
+    sm.KparDiff.pk_dtype = 'float64'
+    sm.KparDist.pk_dtype = 'float64'
     f = SurfaceMatching(Template=ftemp, Target=ftarg, outputDir='../Output/surfaceMatchingTest/'+model +'_'+typeCost,param=sm,
                         testGradient=True, regWeight = regweight, Landmarks = landmarks,
                         #subsampleTargetSize = 500,
